Make the AtomKernel computation parallel
This does come with a slight trade off that the transform method will no
longer be thread safe. This is because it now temporarily stores the other
features on the object to not have to worry about transferring it.

Note: This does mean that all the features will be duplicated among all
processes.

In testing, it seems as though this new formulation is comparable in speed to
the old method when using 1 process, and much faster when working parallel.