Cleanup the graph distance and depth mask code
This correction changes the complexity of the depth mask computation.
Now instead of O(n^(3d)) to compute matrix powers to depth d, it uses
the Floyd-Warshall algorithm which is O(n^3) to compute all graph
distances, then it applies a masking.

By making this change, it corrects a minor bug in the implementation
that caused some atom interaction pairs to get omitted.