Move all encoding code into a separate superclass
This is a fairly drastic change to the internals of all the
encoding-based features. Overall, the change is to make the API a bit
more elegant and to make encoding style features much easier to use in
other contexts. This change allows abstracting the encoding from the
internal representation as a list of values with corresponding buckets.
This change also comes with the nice feature that the general encoding
interface is more consistent between all the different feature types.