Simplfy smoothing functions
This change allows smoothing functions to take a parameter. By doing
this, the one of case of 'circ' is now made to be more sane in the set
of smoothing function options. This change also allows for better
centralization of the smoothing functions so it is now contained in one
area.