Add the ability to apply a function to labels
This is a starting point for creating labels for all features created
with MolML. This is a slightly non-ideal situation because the labels
really should be stored when fitting. Once all the features have labels,
then it should be simpler to go back and make the needed changes for
that.