Change all labels to be ordered
This should be the last in the series of changes to get labels working
in a sane way. The intent of this is to collect all decisions about
whether or not some parts should be sorted to the creation of the label
attributes. This makes things much simpler than before where the sorting
was done when it was needed. This approach should have a lower chance of
bugs, and it makes the code a bit more consistent in how it is written.

Going forward, it is assumed that the LABELS attributes on all
BaseFeatures will be in some standard ordering defined by that
particular feature.