Add a new FormMixin
This is a fairly wide sweeping change. The primary cause for this change
is the disconnect between Encoded* labels and Encoded* features. Before
this change, these would not match up. Secondary to this, is that it
makes the handling of the form information a little bit more explicit
rather than having a collection of objects to pass around.

This change adds a data class (IndexMap) to store the information that
was previously multiple values returned from get_index_mapping. This
replaces the previous collection of primative types that were passed
around and the dynamic function for doing the actual index mapping.

In addition to the approach implemented here, a few other approaches
were considered and were rejected for different reasons. These are
broken down into two main decision points. The first is the structure of
the get_index_mapping data and the second is where these data reside.

Data Structure:
1) Collection of simple objects
    + This is how it was previously done
    + Simpler implementation
    - Error prone
    - The usage of dynamically created functions are potentially
      problematic
    - As previously implemented, this also caused these values to be
      repeatedly recalculated
2) Create a data class to manage data (implemented)
    + Collects related objects into single place
    + Makes it easier to cache/lazy compute these values
    + Allows for a more cohesive/documentable usage
    - Makes saving/loading problematic
        - This is resolved by using a lazy loading interface for the
          actual values

Data Location
1) Add the form information to the EncodedFeature base class
    + For the current usecase, this is simplest
    - This has the problem that the encoding is not dependent on form
    - Conversely, there is no reason to tie form to encoding
2) Add the form information to all implementing classes
    + This is how it was previously done
    - This violates DRY and is error prone
    - This also has potential issues with loading/saving
3) Creating a new mixin (implemented)
    + Abstracts form from other settings
    + Makes it easier to handle the creation/usage of the data
    + More flexible
    - Another mixin to use
    - Potentially a more complex than really needed