Restructure IndexMap to work with multiple outputs
This change is to allow for generalization when selecting which indices
are used for `form` settings. This implementation allows for switching
between the old style of indices and the more complete combinatorial
style.

Additionally, this generlization simplfies some the iterators for encoded
features. This also removes the need for the `both` values as they can now
be handled by having multiple groups.